Only use OVS bridges when container networking method is "provider"
When the container networking method is non-provider (e.g. local or
fan), juju should make use of the setup that the operator has provided
for us instead of forcing the use of an OVS bridge should one be
detected.

On the other hand, if the networking method is set to provider, we can
allow the bridge policy implementation to make use of OVS devices
(subject to any operator constraints) by simply forcing their type to
bridge.