Don't attempt to manually remove lxd veth devices from OVS bridges
The version of lxd that ships with bionic (as an apt package) fails to
automatically remove veth devices if they are attached to an OVS bridge
instead of the default lxd bridge.

In an attempt to address this problem, we introduced extra logic into
the container removal code to effectively run (when the OVS cli tools
are present) "ovs-vsctl --if-exists del-port X" for each veth device.

This command requires to be executed as root and works as expected when
we deploy an lxd workload in an already provisioned machine as the agent
runs as root. On the other hand, if we bootstrap a local lxd cloud, it
is actually the juju CLI that runs the above command which obviously
fails (and bubbles up as an error) as the CLI is not running as root.
Not to mention that the command would have no effect when using a remote
lxd server.

Given that this issue has been fixed in newer versions of juju and there
is not really much we can do to fix this for local/remote lxd clouds,
this commit removes the above cleanup logic.