Export Admin Facade
The following outputs the Admin facade so it makes it possible to login
via another source, other than using the local credentials.

The code is quite simple as we fall back to the facade.Type rather than
asking the registry for the information. We have to do this, as the
Admin facade details is out side of the registry and I'm not sure why
this is...

Fixes: https://bugs.launchpad.net/juju/+bug/1885324