Output a simplified yaml for model-config
The following changes allow us to output a model-config yaml that can
also be consumed by model-config. The type is then isomorphic, as in we
don't loose any information through the usage of this command.

This doesn't break backwards compatibility, but introduces a new output
format: simple-yaml (better name?), that removes the superfluous
information that isn't required by the input for the command.

We could have attempted to parse the original yaml into the simplified
yaml, but we would have issues knowing if it really did conform to the
simplified yaml and would be very error prone.

-----

The command is actually really simple, if you supply `--all` or
`--format=simple-yaml` then it will output a nice neat version of the
yaml. All here, just redirects the previous yaml output to the simple
yaml output, killing two birds with one stone.

-----

The only thing to consider is that `agent-version` is not allowed to be
piped directly into model-config, so is required to be removed. This is
a damn shame and would have been better to have a flag on model-config,
something like --ignore-error.

To get around this, you can do:

```
juju model-config --all > output.yaml && \
    sed -i '/^agent\-version/d' output.yaml && \
    juju model-config ./output.yaml
```