Add new line to juju config output
The following adds the correct standard unix new line after the output
of a config. Originally it was missing, there was some debate about
adding it, but @wallyworld has made the decision to add it.

The code is really simple, just force the usage of fmt.Fprintln, which
will do exactly what operators expect to happen.