Handle new lines in a tabular format.
This change correctly handles newlines when the description contains
them. This is so we can keep tabular formatting and still have the nice
output.

It's not exactly complicated, but just scans over each line and inserts
a tab at the beginning of every line after the first one (it's already
tabbed).