Change how the allwatcher processes unit status updates.
The existing implementation will accidentally overwrite an error status
if a workload status update comes in. Instead of trying to handle both
agent status and workload status updates in one method, made two, one
of which is very trivial.

There was already an existing test for a unit coming out of an error
state, but there was no test for the updating of workload status when
in an error state. That is the test that is added.