Charmhub API find client
This is a very mechanical change for adding the charmhub find client
methods to the existing charmhub client.

The find methods takes a new params.Query instead of a params.Entity as
you might want to utilise search queries that don't align to our strict
names tagging system.

For example, searching for "word press" (with a space) will return
wordpress, but would fail our names.ApplicationTag parsing. Additionally
searching for multiple things "wordpress + mysql" would hopefully be
clever enough to work that out as well.

Essentially we're allowing free form text in the search. We may want to
restrict the amount you can actually search for, but we don't restrict
input anywhere else and to be honest, we'll probably get it wrong.