Consume yaml for model-config
The following changes are a different approach than #11784, where by
that PR attempted to put out a better yaml file for consuming. This PR
attempts to consume the existing yaml. That way we can work with all
sorts of older model-config commands, without having to get operators to
mess about with the various yamls they might have.

The nice thing about this way is that it uses feature detection of the
schema package to detect if a type matches a given format. If that is so
then coerce it into what we expect. This has been employed for the
description package for upgrades and migrations with great success. So
we might as well copy that concept here.

It's really nice, because you can do:

```
juju model-config --format=yaml > output.yaml && \
    sed -i '/^agent\-version/,/source\: .*/d' output.yaml && \
    juju model-config output.yaml
```

 > Unfortunately we have to drop the agent-version as you can't use that
 during a model-config change.

Fixes: https://bugs.launchpad.net/juju/+bug/1879972