Allow reading confing from stdin.
The following changes allow the operator to use normal unix terminology,
with pipes. This simple change makes it super easy to update the
model-config of one model and push to another. We should employ as much
of this as possible as it gives operators more familiarity within the
environment they're used to.