For now rename to charm to entity
This should clear up our API, by renaming charm to entity. We can wait
for the API to update in the future (or not).

With this change, only the client knows or cares about it being a charm
in the API. we can update that at a later date. Within juju though, we
can just talk about entitites.