Adding CharmOrigin when deploying
The following adds charm origin to state. This is independant of the
charm URL, hence why we want to own the origin for the charm. We will
additionally add things like the charm instance id from a charm
resource.

This work is to usher in the new charmhub API. charm.Charm is tied
heavily to the original charmstore, so we want to start modelling the
charm away from the charm.Charm implementation.