Commenting out code due to crazy 3-way merge.
The find code was added while the refactoring of info code was underway
leading to a nasty 3 way merge.  We decided to move forward by
commenting out find code that didn't match as the command is not yet
available for public consumption.  Work to model find data will follow
on, rather than making the info PR larger.