Update find command to the new find API
Pushing the charmhub client API into the API server and reducing how
much the client has to do to show the find/info data for a charm/bundle
makes the client much, much more simpler.

The code changes are really simple, but are laborious, which generally
means it's right. One code becomes complicated you have to wonder if
it's the right implementation.