Remove nbd from lxd profile tests
As our CI machines are not homogeneous we have issues where some of our
tests will work on some machines, but not on others. The very fact that
we have non-homogeneous machines is both an asset and a curse. If they
where all the same, we could ensure that our tests wouldn't be that
flakey. Except that we should be testing the cases where they're not.

In this instance, we're testing LXD profiles and so we need to test the
lowest common denominator and that means removing kernel support from
lxd profiles that we can never meet.

It does beg the question though, how a owner of a charm is expected to
code against this, other than stating that each kernel module is a hard
dependency. This might have been better with feature testing if a module
was available and attempting to use that...