Implement fallback subnet discovery mechanism for older LXD versions
Unfortunately, older LXD versions (e.g. the ones shipping with xenial
and bionic) do not come with the network_state extension enabled. As a
result, subnet discovery for LXD substrates would fail.

This commit introduces a fallback method for reconstructing the subnet
details by introspecting the devices bridged into the controller
container and discovering the host network that they are bridged to.
This method works as subnet discovery is performed after the controller
has been bootstrapped.

The caveat of this method is that we fail to detect any subnet on the
LXD host that is not bridged into the controller container. This is
probably an acceptable tradeoff given that we are always able to detect
the subnet for the lxd bridged (regardless of whether it's managed by
LXD or not).

Finally, juju will emit a warning when using this fallback and suggest
that the operator upgrades their LXD version and runs `juju
reload-spaces` to get full subnet discovery.