Charmhub URL config
This wires up the charmhub url to the model config. It currently makes
a new client every time the method is called, which is not ideal. If
performance does become a problem, we can cache the client for a given
charmhub URL. The operator isn't going to be changing that URL very
often, we should be find having them in a `map[string]Client` type (with
mutex coordination).

The implementation is rather straight forward, with the exception of
adding the charmhub-url config to the immutable slice, prevent further
changes. There is on going discussion about how we correctly model this,
but considering we can ADD mutation at a later date, this PR should be
enough to get us off to the races.