Allow the user to specify formats
Juju has a strong pattern of outputting different formats from each
command to allow the user to script against the CLI. The following
updates both find and info to allow this to happen.

The controversial thing about this change is naming the default output,
as for info the output is some sort of hybrid yaml that isn't yaml and
the default format for find is tabular. To be consistent I've named the
human (naming is hard).

The code is easy other than the warning logger to enable the commands to
output a warning if the result from the API isn't parsable. This can
happen if the charm creator fat fingers something and we don't want to
bring down the find/info commands to the user because of this. We'll
ignore the error (log it out) and present what we have to the user.