View layer
This change includes a *view* layer to the CLI. This doesn't use the API
dto as the view layer, we don't infect the API client with YAML tags and
we don't compromise the API layer with view related things.

Instead we enforce vertical boundary isolation by transforming at EVERY
stage the objects from each part. This is being grown up and owning
every part of the data at every layer.