Simplistic output
The following updates the output from the integration tests to be much
more simplistic. There are now only two modes, the very terse mode,
which is great to just see from a very high level overview if something
works or broke:

   ./main.sh smoke

Or using the verbose mode that aims to be chatty, but not overly chatty
like with the use of `set -eux`. The `x` part makes it hard to see
what's happening and should be only used for local debugging.

   ./main.sh -v smoke

With this simplistic model we should be able to see what's failing
within jenkins, as we use `-v` as the default mode for all our testing.