Have the apiserver watcher code translate unset application status.
If an application status hasn't been explicitly set by a unit its status
is derived from the units themselves. This derivation code was moved into
the model cache, but the all watcher translation later itself was overlooked.

This branch uses the cached controller information to determine the application
status if it is Unset, like the status code does.

There were no tests around the all watcher code on the apiserver side,
so added some to test the conversion behaviour.