Allow the input of a yaml that we output
This follows how model-config works, by allowing users to passing in
what they got, this makes it easier to understand what the command
requires.

The code is very similar to model-config, except model-defaults takes in
to consideration region and controller.

Interestingly we don't consolidate when default and controller are the
same value, which seems a shame.