Identification code for empty API Addresses.
Users in discourse have identified a situation where juju agents have
empty api addresses written out to their agent.conf files. It's not yet
clear on why this happening but we believe that the api address update
worker is the one responsible for writing out the empty values to the
agent.conf file.

This change introduces some logging and a conditional statement to not
allow writing out of empty api address structures.

LP: 1888453
Discourse: https://discourse.juju.is/t/file-agent-conf-missing-apiaddresses-entry-after-network-outage-or-upgrade/2975