Ensure that we validate model-defaults
Model-defaults doesn't validate the configuration that you pass to it,
so you can end up shooting yourself in the foot at a later stage. They
way we validate here is slightly different from model-config. The code
checks against the current model for validating against the schema. It
does not however take in the resulting model config, instead it uses the
origin because it passed the validation aspect.