Import the new description changes
This wires up the new changes, but doesn't finish wiring them up.
Instead it joins up the new charm origin, without parsing at every
stage. That will come later when we implement the download, refresh API.

Part of these changes include the parsing and typing of a charm
channel. It conforms to the snapcraft channels docs[1] and collapses
where possible to be consistent with snap.

Both migration import and export is wired up, but again these are place
holders for when more data might be needed.

1. https://snapcraft.io/docs/channels