Introduce a nested context for the deployer worker.
This isn't yet hooked up into the deployer, but the nested context will
be hooked up to have the unit workers run in a nested dependency engine within
the deployer worker.

The Context interface is expanded to support the worker.Worker methods as
the NestedContext is a worker, but the SimpleContext has nothing running,
so its implementation are no-ops.