Refactor open port manipulation code to work in an endpoint-aware manner
This commit brings a series of significant changes to the way that juju
represents per-machine open port ranges. To begin with, instead of
having a per-machine, per-subnet document in the the "openedPorts" collection,
we now use a *single* per-machine document which maintains a list of
sub-documents that keep track of the open port ranges for each unit on
the machine grouped by application endpoint.

The endpoint to port ranges (a list) map may also include an empty ("")
key which acts as a wildcard to indicate that a set of port ranges has
been explicitly opened by the charm for *all* endpoints. By recording
this bit of information, juju will be able to decide whether to open
these port ranges to any new endpoints that might appear after upgrading
a charm.

As far as document ID conventions are concerned, the opened ports
collection will now use the machine ID as the doucment ID (instead of
"machineID#subnetID") as there is a one-to-one mapping of machines to
opened ports documents.

This new way of representing port ranges will enable juju to lazily
resolve each port range to a subnet where needed (e.g. in the
firewaller). Furthermore, as opposed to the approach of resolving
endpoint bindings to subnets and storing those, this new approach
requires much less work to keep everything in sync when either the space
topology changes, or when the operator changes the bindings for the
deployed applications.

The changes in this commit introduce an improved API for querying and/or
manipulating opened machine ports. The API consists of two interfaces:
- MachinePortRanges: methods for accessing the opened ports for a
  machine, querying the unit set of open port ranges (across all subnets) and
  for obtaining a ModelOperation to apply any mutations to the opened
  port ranges.
- UnitPortRanges: methods for querying the opened ports for the unit and
  for queuing requests for opening/closing ports for a particular
  endpoint.

Implementation-wise, the commit provides the machinePortRanges type (a
concrete MachinePortRanges implementation) which encapsulates a document
from the openedPorts collection and unitPortRanges (a UnitPortRanges)
implementation that is nothing more than a view on the machinePortRanges
instance. Any change requests on a unitPortRanges type are recorded to
the underlying macinePortRanges instance and can be committed in a single
txn.

The txn-building logic for opening/closing port ranges has been revamped
to cope with scenarios such as:
- Open the same port range for the same unit but different endpoints
- Open port range for a specific endpoint while the unit has also opened
  the same range for all endpoints.
- Open port range for all endpoints and close it for a particular
  endpoint. In this case, the entry is removed from the wildcard
  endpoint and replaced with individual entries for all remaining
  application endpoints.

Finally, this commit ensures that all tests in the state package are
green after this set of changes. So as to avoid having to also change
the facade code in the same commit (some tests in the state package spin
up a full API server), the legacy open port API method bodies
(unit/machine) that used to invoke code that has been removed from the
state package have been replaced with TODO panic messages which will be
removed via the following commits.