Some updates from the review.
A key change is that the failure to start a unit is no longer
fatal to the deployer. And ErrTerminateAgent will result in a unit
being stopped, and ErrRestartAgent will just bounce the dependency
engine for the unit.

A simple hub is introduced here for async calls. While not entirely
necessary for this piece of work, the hub is used in an upcoming
branch were we introduce the ability for the introspection endpoint
to stop and start units.