Update firewaller worker/facade code to use the new port range API
Even though in pre-2.9 juju installations the controller and firewall
worker includes support for ports on a per-subnet basis, this
functionality is not exposed via the facade API. Instead, ports are
opened in all subnets (signified by the use of a document with an empty
subnet ID).

To this end, this commit updates the firewall worker and related facades
to emulate this behavior using the new {Machine,Unit}PortRanges types.
For simplicity, the firewall facades that currently *do* accept a subnet
ID will return a not-supported error if one is provided. This is
generally safe as this functionality is triggered by watching the open
ports collection which only includes docs with an empty subnet ID (as
limited by the current implementation).

In the future, we will be introducing a new API method for obtaining the
list of port ranges opened by each unit (for a particular machine) in
each subnet (as dictated by the space bindings for each unit's
application). Then, we can complete the refactoring work that will allow
the firewaller to work with endpoints.