Improve user experience with network settings.
If no network is specified in the config, attempt to look for a single
internal network. If one is found, use it.  If more than one is found,
error with instructional message.  If no network is found, attempt to
create the instance anyway.  There are clouds like this.

If we fail to start instance with a BadRequest related to no defined
network, error with an instructional message.