Add Changes method to the UnitPortRanges interface
This commit allows the machine port operations implementation to filter
pending changes by a particular unit name. This change allows us to
expose a Changes() method to the UnitPortRanges for obtaining a
ModelOperation scoped to the changes for a particular unit only.

With this change, the Unit.OpenedPortRanges does not have to return a
"change function" any more.