Remove the unknown charmorigin source type
The charmorigin source type; unknown is surpless to requirements. There
is only one place that is left and that is in state. On importing
migrations the tests expect that there is no charm origin and so cause
all the tests to fail if we expect there to be one.

The fix here is to ensure that we handle unknown further up the stack
and attempt to deduce it from the charm url if it's unknown.