Add message to no network instance start failure.
Adding a single found internal openstack network to the model's config
would be ideal. Unable to do so correctly for model's other than the
controller at this time.  Add an error message to a failure due to
not configured error.

There are some OpenStack configuration where no network is available,
and instances still start correctly.  Therefore we cannot assume that no
configured network is a failure.