Allow the upgrading of existing applications
This adds an upgrade step to allow for given applications to have a
charm origin. There is a real fundamental problem and that is that the
charm url with in applications must have a schema. You can't have a
schema-less charm url, it throws errors when you try and parse and
unmarshal it.

SO at some point we'll need to update the charm package to allow for
either schema less endpoints OR to have a schema charm hub url. Either
one sounds bad.