Hook up the deployer nested context.
This work introduces the implementation of the unified machine
and unit agent for IAAS models. The deployer now runs the workers
for the units in the same process using a runner to run a primary
worker for each unit. This primary worker is a dependency engine
that runs a subset of the workers that the old unit agent ran.
There is no need for the units to now run upgrade workers as
that is all handled by the machine agent.

An upgrade step is added to disable all the service files running
for the old unit agents, and update the agent.conf file for the
machine agent to record the deployed units.