Update CommitHookChanges implementation to accept endpoints
The key change here is that the Open/ClosePort payload for the
CommitHookChanges call needed to be augmented so that it includes an
endpoint field.

The uniter facade version has been bumped to V17 to signal this change.
The change is backwards compatible with older clients as an empty
endpoint value indicates that the open/close operation applies for all
endpoints thus matching the pre-2.9 behavior.