Introduce bash functions to stop and start unit workers.
The introspection endpoint gains a new endpoint to handle post
methods to request unit start and stop. This endpoint then uses
a new simple hub to publish the request.

The deployer now requires a simple hub to be passed in through the
manifold rather than creating one itself. The internal stop unit
message is now replaced with a centrally defined one. An additional
start unit method, as well as a status method, is added.

The scripts file that is written out into /etc/profile.d is updated
to remove the unit agent aspects, as now there is only a machine
agent. This also allows better argument passing from the other
defined functions.