Wire new APIs to uniter and port-related hook tools
In addition, this commit removes all the redundant port range validation
code (and tests) as port parsing is delegated to core/network
ParsePortRange which is thoroughly unit tested.

While the hook tools use the new API methods, for the time being, they
stil pass an empty endpoint to the relevant hook context methods
(Open/ClosePortRange). This will be addressed in the following commits.