Implement controller-side of the OpenedMachinePortRanges firewaller call
The call has the same signature with the similarly named call of the
Uniter (V17) facade and shares the same response payload. However, while
the uniter cares about the port ranges opened by units grouped by
endpoint, the firewall worker requires the port ranges grouped by a
subnet CIDR.

This entails a convoluted mapping step where we make use of the
following bits of information:
- the application endpoint to spaceID bindings for each machine unit
- the space ID to endpoint CIDRs mappings for the targeted machine

Using this information, we can transitively resolve an endpoint (port
ranges are stored in the DB grouped by unit and endpoint name) to a list
of subnets and group the port ranges by CIDR.

Once this call is wired to the firewaller, we will be able to lazily
resolve the opened port ranges into the CIDRs needed for generating the
firewall ingress rules.