Replace ifNotMigrating with ifFullyUpgraded for deployer manifold.
When the unit agents were moved to run inside the machine agent,
migration started to fail.  The deployer worker was being stopped during
the migration quiesce phase due to ifNotMigrating, thus stopping the unit
agents. However they are required to be running during migration.

The deployer still needs ifFullyUpgraded, which was wrapping the
migration fortress.  Add it, as we do not want to deploy during
upgrades.