Use SpaceInfos/AllEndpointBindings to optimize OpenedMachinePortRanges
As pointed out by the code reviewers, it would be much more efficient to
use SpaceInfos for creating the spaceID -> []subnet CIDR mappings (the
previous implementation in state/machine would iterate the addresses and
make a query for each one to grab the subnet).

In addition, the SpaceInfos can be lazily fetched when needed (we would
need at least one valid machine after performing the ACL checks to do
so) and re-used for all other entities present in the OpenedMachinePortRanges
call.

The same observation applies to obtaining the endpoint bindings for the
applications deployed to a particular machine. A better alternative is
to fetch (and memoize) the bindings for all model applications. Note
that this approach will scale even to large models as the number of
deployed applications will be << deployed units.