Add juju_leases introspection command.
This branch adds a /leases endpoint in the introspection http handler.
The new juju_leases function wraps up calls to this function and allows
for the user to easily add application names (or partial names) or
part of a model UUID using -m.

The result is always formatted YAML. This conveniently sorts all the
keys for us, and would allow other tools to programaticly interrogate
the endpoint if need be.