Remove incorrect assumption.  Shorten goose error message.
The assumption you cannot create an OpenStack instance with no network
defined if there is an vailable network is false.  Remove that logic.

Keep the output change to suggest a network to specify if
tryStartInstance fails due to missing network.

If StartInstance fails due to error from RunServer, print the cause, not the
entire thing which includes a []byte.  Instead print the error cause,
and log the entire message as Debug.