Update the charm dependencies
The following is a very mechanical change for updating the go charm
dependencies, to ensure we have the latest and greatest stuff.

Incidently, I do believe we're doing this wrong. The idea that we have
multiple repos that all self depend on each other is slightly sad. We
should instead internalise both charmrepo and bundlechanges into Juju
directly.