Facade schema groups
This introduces facade schema groups to the schema gen. It defines these
restricted groups so that it makes it easy to understand what the schema
gen generates.

There are 4 groups:

 1. latest: this is the default group and the one that was previously
 running.
 2. all: this outputs all the facades, everything!
 3. client: this outputs just the facades for client interactions and
 probably should actually be the default!
 4. jimm: this is the JIMM facades.

We can tailor each group so that they're defined and restricted to a
given release. We could make all the facades a giant list with versions,
but then it becomes harder to understand because of the many versions.