Target latest mongo
The client tests rely on a certain mongo installed for github.
Unfortunately github actions install their own default. We can't have
that, as we require what's packaged with the apt cache for the OS (there
is better verbiage for that statement!).

Anyway we brute force our way through the mongo versions and remove them
from the source list. THe code is terrible and annoying at the same
time. I'm still unsure how we requested ubuntu and we ended up with a
bunch of software on it that I didn't ask for!