Initial idea for a charm procedure
The idea behind this code is to have a procedure to run the code that is
found in AddCharmWithAuthorizationAndRepo in a more generic and
standardised way.

The code is very helpful for charmstore and charmhub, so making it more
generic is highly desirable.

This is more of a WIP atm, so more changes are to be expected.