Rename procedure to strategy.
Adding to tests to solidify the idea of a download strategy. We should
be able to run this code in multiple places, as we've totally abstracted
out the idea of what the input and the output of the strategy is for
downloading a charm from a given store.

The code still needs integration with the existing application
charmstore code, but for now it works the way we want.