Use the correct LXDProfile
The LXDProfile state is different that the LXDProfile from core. The
apiserver core is at the rough point where we're using both state and
core LXDProfiles.

This is a shame, we should have clear boundaries where we transform a
LXDProfile from one to another.

The code ensure that we have the correct LXDProfile for each scenario at
the right time, it's just a bit convoluted.