Update CharmOrigin in api and params.
Add Risk and change Channel to Track for the client side and going over
the wire.  A CharmStore channel is a risk according to CharmHub.  This
gets complicated if a single string pointer is going over the wire to
represent both types.  Keep separate until the core charm channel can
take over.

Moved the CharmOrigin from both application and charms on the client
side to a single location.  This will prevent translation issues between
the two in the juju client.