Upgrade-series validation
The following updates the upgrade-series validation to ensure that you
haven't already got a prepare already running for that machine. Although
it is eventually picked up when trying to create a lock, the user had to
jump through a series of prompts to find that information out.

Instead we just lift the validation higher and the problem surfaces much
sooner in the work flow.