Fix intermittent fail in TestBuildModelRepresentationApplicationsWithSubordinate.
Sort slices of names in BuildModelRepresention to ensure tests return
good results for application config etc.

Gomock expect does not handle "...string" as an arg well.  If the
strings provided to the reall method call are ("wordpress", "sub) and
the gomock EXPECT is for ("wordpress") it will believe the method is
called.  If the slice gets reversed to ("sub", "wordpress), the same
EXPECT will fail because "sub" does not equal "wordpress".  In the first
case, "sub" is ignored.