Provide upgrade-series status information
The following provides status information for when an upgrade series
fails the validation check. Additionally we provide an optional typed
set of information that we can pass with the error. This info in the
error contains the status, this can then help the user out about what
to do next.

We do it this way for 2 reasons:

 1. The API facade doesn't need to bumped, this is purely optional error
 message information. The status is no critical, if the information is
 malformed or wrong, we always show the underlying error message.
 2. We don't need to lift the error messages from the client to the api
 server, as we can switch on the type. This gives us more control about
 what we want to show at the client level without making the server
 error messages long and verbose.

This maybe a dark art of API server error messages, but I'm starting to
think this might be a good way to offer optional state hints to the user
without confusing the user with cryptic error messages.