Validate the machines under a model
The following code validates a machine under a given model, but ignores
controller models.

The code handles connecting to the state pool, to get a model that was
requested and then ensuring that there isn't a machine under going a
upgrade series.

One downside to doing this way, is that if there are any models with
1000s of machines, this could be really slow. Unfortunately there isn't
really much we can do at this stage. Maybe in subsequent commit we can
make a bulk method for checking if any machines in a model are
upgrading...