Ensure we test that a upgrade mode cmd works
The following ensures that when there is an error for verification of an
upgrading model, we correctly error out without continuning.

The mocking and faking of this command is a bit unwieldy and because of
that it's hard to full do unit level tests because of that. Anyway I
believe we're at a level that makes most sense.