Revise response payload for OpenedMachinePortRanges API call
This response payload was recently introduced to support the open-ports
for endpoints work and would initially group the port ranges for each
unit by either endpoint name or resolved subnet CIDR (for the bound
space) depending on whether the response was generated by the uniter
facade or the firewaller facade.

However, for the expose application endpoints work, the firewaller must
be able to access this information using *both* grouping modes. To this
end, the payload had to be modified so that it can include multiple
groupings.