Use dedicated payload for uniter's open machine port ranges API
Prior to this commit, the OpenedMachinePortRanges call for the uniter
and the firewaller facades used the same response payload. With this
change, the uniter call has been renamed to
OpenedMachinePortRangesByEndpoint to better indicate its purpose and it
now has its own dedicated payload type with one less level of
indirection.