Use dedicated payload for firewaller's OpenedMachinePortRanges API
In order to implement the expose application to endpoint feature, the
firewaller needs to be aware of not only the opened port ranges by
endpoint for each unit but also of the subnet CIDRs that correspond to
the spaces bound to each endpoint.

To achieve this, the aforementioned firewaller API now uses its own
dedicated payload type which is quite similar to the
OpenedMachinePortRangesByEndpoint uniter API but additionally includes
the resolved CIDRs for each endpoint.