Always show app status message; rename app Notes column to Message
Previously the application status message was only shown for CAAS
models, but we'd like to show it all the time for consistency and to
address the issue in this bug.

This also removes display of "exposed" (which was overwritten when a
message was present in any case) as that is being revamped as we speak.

Updated go-check to the latest version, which pulls in this minor
documentation fix:
https://github.com/go-check/check/commit/038fdea0a05bc030b0bfda479dc2e08d2220ec74

Bug link: https://bugs.launchpad.net/juju/+bug/1850268