Rework the download charm client
The previous download client was a bit to close to what the charm repo
did. It didn't do what we wanted it to do and wasn't what we expected.
So I clean out most of the code, gave it a better set of interfaces,
ones which are testable and away we went.

Along the way I renamed a lot to make more sense, but tbh, there wasn't
much wrong!