Increase AWS SDK retries from 3 to 10; fetch 1000 "offerings" at once
Retries are turned on by default with the AWS SDK, but they're very
light: 3 retries with an initial 500ms delay and doubling backoff (so
500ms, 1s, 2s, give up). Whereas the retries in the amz library (in
EC2.query) are much heavier: 10 total attempts with an initial delay of
1s and doubling backoff (with max 1 minute, so 1s, 2, 4, 8, 16, 32, 60,
60, 60 -- about 4 minutes total). I've bumped up the AWS SDK retries to
roughly match the EC2.query ones. This fixes (or at least masks)
https://bugs.launchpad.net/juju/+bug/1888409

We were doing 39 AWS API requests per InstanceTypes call: 1 for zones,
16 for "offerings", 16 for types, 6 for spot price. The types and spot
price are already on maximum # of results per page, but I've increased
the "offerings" one from 100 to 1000. That means: 1 zones, 2 offerings,
16 types, 6 spot price = 25 total = 36% reduction in API calls.