Change the default stderr logging to WARNING
We have been logging both stdout and stderr for units at DEBUG, but that
leaves everything at debug. And while we want to stop defaulting to
including DEBUG, we want to default to include things like Tracebacks,
which come from stderr.
So split the logging, and log all of stdout to DEBUG and stderr to
WARNING.