Change the behavior of SetExposed method to merge and rename it
The operator is allowed to perform multiple "juju expose" invocations
(e.g. one for each endpoint) and the intent is to accumulate expose rules
instead of simply replacing them.

To this end, the SetExposed method of the application model in state has
been renamed to MergeExposeSettings and updated to merge the incoming
changes into the current set of exposed rules.