Support handling of expose parameters in SetExposed API call
The existing payload has been augmented with an extra field for
specifying the per-endpoint expose parameters (expose to
{spaces,CIDRs}). The new field will be used by compatible clients to
expose applications in a more granular fashion.

If the new field is not populated, the call is equivalent to the pre 2.9
expose call where *all* endpoints are exposed to 0.0.0.0/0.
Consequently, this change is backwards compatible. However, compatible
clients should perform the usual facade version checks to prevent calls
to SetExpose when the per-endpoint parameters are specified and a
pre-2.9 controller is targeted.