Support handling of endpoint lists in Unexpose API call
The existing payload has been augmented with an extra field for
specifying an optional list of endpoints to unexpose. If the list is
populated, only those endpoints will be unexposed and the application
will remain exposed if other endpoints remain exposed. On the other
hand, if the list is empty the application will be unexposed.

The latter ensures that we remain consistent with pre 2.9 clients.